requirements.txt
pandas~=2.3.1
scikit-learn~=1.6.1
joblib~=1.5.1
matplotlib~=3.10.3
seaborn~=0.13.2
python-docx
pytest
pyyaml~=6.0.2
openpyxl
imblearn~=0.0
scipy~=1.16.0
numpy~=2.3.1